Struct ncomm_executors::simple_executor::SimpleExecutor

pub struct SimpleExecutor<ID: PartialEq> { /* private fields */ }

Simple Executor

This simple executor stores Nodes in a sorted vector where the priority is higher the closer to the current timestamp the Node’s next update is.

Note: The Simple Executor can be interrupted by sending a true value over the mpsc channel whose receiving end is owned by the SimpleExecutor

Addendum: The Simple Executor will also busy wait between node executions so do not expect the SimpleExecutor to yield CPU time to other processes while it is running.

Implementations

impl<ID: PartialEq> SimpleExecutor<ID>

pub fn new(interrupt: Receiver<bool>) -> Self

Create a new Simple Executor without any Nodes

pub fn new_with( interrupt: Receiver<bool>, nodes: Vec<Box<dyn Node<ID>>>, ) -> Self

Creates a new Simple Executor with a number of Nodes